+++ react/CHANGELOG.md
@@ -17,12 +17,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [10.0.1] - 2020-09-02
+
+### Fixed
+
 * Fix rendering of failed hooks from Cucumber-JVM
   ([#1166](https://github.com/cucumber/cucumber/issues/1166)
    [#1167](https://github.com/cucumber/cucumber/pull/1167)
    @aslakhellesoy
    @sebrose
    @cbliard)
+* Fix react warnings about rendering `<li>` inside `<li>`
+* Updated some class names
 
 ## [10.0.0] - 2020-08-08
 
@@ -246,7 +252,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v10.0.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v10.0.1...master
+[10.0.1]:      https://github.com/cucumber/cucumber/compare/react/v10.0.0...react/v10.0.1
 [10.0.0]:      https://github.com/cucumber/cucumber/compare/react/v9.0.0...react/v10.0.0
 [9.0.0]:      https://github.com/cucumber/cucumber/compare/react/v8.2.0...react/v9.0.0
 [8.2.0]:      https://github.com/cucumber/cucumber/compare/react/v8.1.0...react/v8.2.0